Move composition local to `ResellTheme`
