+++ app/build.gradle.kts
@@ -151,7 +151,6 @@ dependencies {
     implementation(libs.ui)
     implementation(libs.ui.tooling.preview)
     implementation(libs.ui.tooling)
-    implementation(libs.compose.shimmer)
 
     implementation(libs.kotlinx.coroutines.core)
