+++ app/src/main/java/com/cornellappdev/resell/android/ui/components/global/ResellCard.kt
@@ -36,10 +36,8 @@ import com.cornellappdev.resell.android.ui.theme.ResellPreview
 import com.cornellappdev.resell.android.ui.theme.Secondary
 import com.cornellappdev.resell.android.ui.theme.Stroke
 import com.cornellappdev.resell.android.ui.theme.Style
+import com.cornellappdev.resell.android.util.shimmer
 import com.cornellappdev.resell.android.viewmodel.ResellViewModel
-import com.valentinilk.shimmer.ShimmerBounds
-import com.valentinilk.shimmer.rememberShimmer
-import com.valentinilk.shimmer.shimmer
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
@@ -114,21 +112,15 @@ private fun ResellCardContent(
 
 @Composable
 private fun AnimatedClampedAsyncImage(image: ResellApiResponse<ImageBitmap>) {
-    val shimmer = rememberShimmer(
-        shimmerBounds = ShimmerBounds.Window
-    )
     Box(modifier = Modifier.heightIn(min = 130.dp, max = 220.dp)) {
         AnimatedContent(targetState = image, label = "image loading") { response ->
             when (response) {
                 ResellApiResponse.Pending -> {
                     Box(
                         modifier = Modifier
-                            .shimmer(shimmer)
                             .height(175.dp)
+                            .shimmer()
                             .fillMaxWidth()
-                            .background(
-                                Secondary
-                            )
                     )
                 }
