+++ app/src/main/java/com/cornellappdev/resell/android/ui/theme/Theme.kt
@@ -1,13 +1,19 @@
 package com.cornellappdev.resell.android.ui.theme
 
 import android.os.Build
+import androidx.compose.animation.core.InfiniteRepeatableSpec
+import androidx.compose.animation.core.animateFloat
+import androidx.compose.animation.core.keyframes
+import androidx.compose.animation.core.rememberInfiniteTransition
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.dynamicDarkColorScheme
 import androidx.compose.material3.dynamicLightColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.platform.LocalContext
+import com.cornellappdev.resell.android.util.LocalInfiniteShimmer
 
 private val DarkColorScheme = darkColorScheme(
     primary = Purple80,
@@ -38,9 +44,35 @@ fun ResellTheme(
         else -> LightColorScheme
     }
 
-    MaterialTheme(
-        colorScheme = colorScheme,
-        typography = Typography,
-        content = content
-    )
+    // Create an infinite transition for animation
+    val transition = rememberInfiniteTransition()
+
+    // Animate a value from 0 to 1 infinitely
+    val animatedColor = transition.animateFloat(
+        initialValue = 0f,
+        targetValue = 1f,
+        animationSpec = InfiniteRepeatableSpec(
+            animation = keyframes {
+                durationMillis = 2000
+                0f at 0
+                1f at 1000
+                0f at 2000
+            }
+        ),
+        label = "infinite loading"
+    ).value.let {
+        interpolateColorHSV(
+            startColor = Wash,
+            endColor = Stroke,
+            fraction = it
+        )
+    }
+
+    CompositionLocalProvider(LocalInfiniteShimmer provides animatedColor) {
+        MaterialTheme(
+            colorScheme = colorScheme,
+            typography = Typography,
+            content = content
+        )
+    }
 }
